display a changelog using a range of commits
Adds a new option 'rev-list'.
rev-list lets you define which commit to start from when displaying the
changelog. You can use a tag, a branch, a commit hash, an explicit
range, or anything else supported by git-rev-parse. Examples:

.. git_changelog::
    :rev-list: v1.0..HEAD

.. git_changelog::
    :rev-list: master..topicbranch

Consult the man pages of git-rev-parse for more details on the syntax.