Add git_commit_detail directive.
This directive inserts information about the state of the repository
when sphinx is run to generate documentation.
The following options are available:
branch -- show the name of the branch
commit -- show the commit hash
sha_length -- number of (hex) characters to show for the commit
uncommitted -- show a warning if there are uncommited changes
untracked -- show a warning if there are untracked files

Update documentation to include git_commit_detail.
Add link to GitHub commit, with option to disable, and update the doc.